pyproject.toml
[project]
authors = [
    { name = "Slava Chaunin", email = "67190162+Ssslakter@users.noreply.github.com" },
]
dependencies = [
    "torch>=2.8,<2.9",
    "numpy",
    "datasets>=4.4.1,<5",
    "matplotlib",
    "transformers>=4.57.3,<5",
    "jupyter",
    "tqdm",
    "trackio",
    "torchvision",
    "hydra-core>=1.3.2,<2",
]
name = "trainer-tools"
requires-python = ">= 3.11, <3.14"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
trainer-tools = { path = ".", editable = true }


[tool.pixi.tasks]


[tool.black]
line-length = 120
